requirements.txt
Flask
sentence-transformers
torch
PyPDF2
python-docx